Jwt problem

See merge request s08-webmobile1-sub2/S08P12A706!68

+++ back-end/liargame/src/main/java/com/sixsense/liargame/api/service/impl/CommentServiceImpl.java
@@ -1,53 +1,57 @@
 package com.sixsense.liargame.api.service.impl;
 
-import com.sixsense.liargame.api.request.CommentReq;
 import com.sixsense.liargame.api.response.CommentResp;
 import com.sixsense.liargame.api.service.CommentService;
 import com.sixsense.liargame.db.entity.Comment;
 import com.sixsense.liargame.db.repository.CommentRepository;
-import com.sixsense.liargame.db.repository.UserRepository;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
 public class CommentServiceImpl implements CommentService {
     private final CommentRepository commentRepository;
-    private final UserRepository userRepository;
-    public CommentServiceImpl(CommentRepository commentRepository, UserRepository userRepository) {
+
+    public CommentServiceImpl(CommentRepository commentRepository) {
         this.commentRepository = commentRepository;
-        this.userRepository = userRepository;
     }
 
     //댓글 생성
     @Override
-    public Comment insertComment(Long articleId, CommentReq commentReq) {
-        Comment comment = commentRepository.save(commentReq.commentToEntity(articleId));
-        return comment;
+    public void insertComment(Long userId, Long articleId, String content) {
+        commentRepository.save(toEntity(userId, articleId, content));
     }
 
     //댓글 삭제
     @Override
-    public void deleteComment(Long commentId) {
-        commentRepository.deleteById(commentId);
+    public void deleteComment(Long userId, Long commentId) {
+        Comment comment = commentRepository.findById(commentId).orElseThrow();
+        if (Objects.equals(comment.getUserId(), userId))
+            commentRepository.deleteById(commentId);
     }
 
     //댓글 업데이트
     @Override
-    public Long updateComment(Long articleId, Long commentId, CommentReq commentReq){
+    @Transactional
+    public void updateComment(Long userId, Long commentId, String content) {
         Comment comment = commentRepository.findById(commentId).orElseThrow();
-        comment.updateComment(commentReq.getCommentContent());
-        commentRepository.save(comment);
-        return commentId;
+        if (Objects.equals(comment.getUserId(), userId))
+            comment.updateComment(content);
     }
 
     //게시글의 댓글 전부 가져오기
     @Override
-    public List<CommentResp> findAllComments(Long articleId) {
+    public List<CommentResp> findAllComments(Long articleId, Pageable pageable) {
         Sort sort = Sort.by(Sort.Direction.DESC, "updatedAt");
-        List<CommentResp> comments = commentRepository.findAllByArticleId(articleId, sort);
-        return comments;
+        PageRequest pageRequest = (PageRequest) pageable;
+        pageRequest.withSort(sort);
+        List<Comment> comments = commentRepository.findAllByArticleId(articleId, pageRequest);
+        return comments.stream().map(CommentResp::new).collect(Collectors.toList());
     }
 }
